Share the start option scanner between parser and engine
